feat: InputField 컴포넌트 common 폴더로 파일 경로 변경
